MIT license file to encourage open source use
Hi, Dan! I liked your blog article (https://blog.bitlabstudio.com/a-simple-chat-app-with-react-node-and-websocket-35d3c9835807) and since it sounds like you are open to building upon this, I thought I would propose adding an open source license. Up to you, of course, if you would prefer a different license! (If you wanted to do that, instructions here: https://docs.github.com/en/github/building-a-strong-community/adding-a-license-to-a-repository)